Release 3.0
Significant reworking:

slackscan now perfoms the filter, blacklist and protect file processing
that was previously left to slackup. The result is that slacscan
now uses 'configuration profiles' instead of a config file.

slackup is now included in this distribution as it no longer makes
sense to ship it as a separate project given that so much of the
functionality is now in slackscan.