slackscan: Handle duplicate packages in single source location
Pat recently added a new kernel package to patches/packages/ while leaving
the old ones in there as a fallback. This caused slackscan to fail to
detect the availability of updated packages.

To handle this in future add a reverse version sort on the full package name,
followed by a unique sort on the short package name when scanning for available
packages.

This should result in only the higher version of any duplicate packages
being scanned, assuming that the version numbering scheme follows one
recognised by sort -V.