remove sample kernel-modules post-install script
Slackware post 15.0 reworked the kernel packages/management and no
longer ships a kernel-modules package.

Also, post-15.0 'current' added the geninitrd framework for doing
post kernel install processing. While users can disable that feature
and carry on using slackup post-install scripts for this purpose,
I don't want to add any additional confusion to the situation, so the
kernel post-install will be left as an exercise for the user, letting
them choose to go the stock 'geninitrd' way, or not.