slackscan: add new option to generate a protect file
New option '-S' will cause slackscan to generate a new protect
file containing a pattern for each package currently installed
on the system. This can act as a safety net to prevent
accidental removals should package files unexpectedly become
unavailable.

As it is, slackscan will only generate removal lists if you ask
it to and slackup already has the -R option to exclude removals,
but this feature provides another mechanism to prevent disaster
and as the other two options are all or nothing this provides a
more granular approach as the autogenerated protect file can be
edited by the user after creation.

It will likely be a niche option that most users won't need.