slackscan: apply filter and blacklist on a per package-source basis
filter and blacklist were applied on the combined results of a scan,
however this meant that the user had no control over the filtering
of specific versions of duplicate packages in a single source
location: highest version always won.

By filtering each location in turn we allow the user to explicitly
exclude the higher versioned duplicate before the built-in version
based duplicate resolution occurs.

As well as providing more control, this should be more inline with
user expectations, as it was previously possible for duplicate
resolution to exclude the lower version of a package only to have
the user exclude the higher version using a filter/blacklist,
leaving no version at all.